+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 -   Update ligo-gracedb to 2.0.0.dev0.
 
+-   Add superevent duration for gstlal-spiir pipeline.
+
 ## 0.0.22 (2018-07-11)
 
 -   Process gstlal-spiir events.